Add sidebar component
