+++ app/javascript/components/Home.jsx
@@ -1,13 +1,4 @@
 import React, { Fragment, useState } from "react";
-import { Dialog, Menu, Transition } from "@headlessui/react";
-import {
-  Cog6ToothIcon,
-  ShoppingCartIcon,
-  CircleStackIcon,
-  PlusIcon,
-  SignalIcon,
-  XMarkIcon,
-} from "@heroicons/react/24/outline";
 import {
   Bars3Icon,
   ChevronRightIcon,
@@ -15,277 +6,48 @@ import {
   MagnifyingGlassIcon,
 } from "@heroicons/react/20/solid";
 import Logo from "./Logo";
-import RecentActivity from "./RecentActivity";
 
-const navigation = [
-  {
-    name: "Fishing Rigs",
-    icon: CircleStackIcon,
-    current: false,
-    href: "/rigs",
-  },
-  {
-    name: "Rig Parts",
-    icon: CircleStackIcon,
-    current: false,
-    href: "/fishing_components",
-  },
-  {
-    name: "Gear Reccomendations",
-    href: "#",
-    icon: ShoppingCartIcon,
-    current: false,
-  },
-  { name: "Activity", href: "#", icon: SignalIcon, current: false },
-  { name: "Settings", href: "#", icon: Cog6ToothIcon, current: false },
-];
-const statuses = {
-  offline: "text-gray-500 bg-gray-100/10",
-  online: "text-green-400 bg-green-400/10",
-  error: "text-rose-400 bg-rose-400/10",
-};
 const environments = {
   Preview: "text-gray-400 bg-gray-400/10 ring-gray-400/20",
   Production: "text-indigo-400 bg-indigo-400/10 ring-indigo-400/30",
 };
-const activityItems = [
-  {
-    user: {
-      name: "Michael Foster",
-      imageUrl:
-        "https://images.unsplash.com/photo-1519244703995-f4e0f30006d5?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
-    },
-    projectName: "carolina rig",
-    branch: "freshwater",
-    date: "1h",
-    dateTime: "2023-01-23T11:00",
-  },
-];
-
-function classNames(...classes) {
-  return classes.filter(Boolean).join(" ");
-}
 
 export default function Home({ activity }) {
-  const [sidebarOpen, setSidebarOpen] = useState(false);
-
   return (
     <>
-      <div>
-        <Transition.Root show={sidebarOpen} as={Fragment}>
-          <Dialog
-            as="div"
-            className="relative z-50 xl:hidden"
-            onClose={setSidebarOpen}
-          >
-            <Transition.Child
-              as={Fragment}
-              enter="transition-opacity ease-linear duration-300"
-              enterFrom="opacity-0"
-              enterTo="opacity-100"
-              leave="transition-opacity ease-linear duration-300"
-              leaveFrom="opacity-100"
-              leaveTo="opacity-0"
-            >
-              <div className="fixed inset-0 bg-gray-900/80" />
-            </Transition.Child>
-
-            <div className="fixed inset-0 flex">
-              <Transition.Child
-                as={Fragment}
-                enter="transition ease-in-out duration-300 transform"
-                enterFrom="-translate-x-full"
-                enterTo="translate-x-0"
-                leave="transition ease-in-out duration-300 transform"
-                leaveFrom="translate-x-0"
-                leaveTo="-translate-x-full"
-              >
-                <Dialog.Panel className="relative mr-16 flex w-full max-w-xs flex-1">
-                  <Transition.Child
-                    as={Fragment}
-                    enter="ease-in-out duration-300"
-                    enterFrom="opacity-0"
-                    enterTo="opacity-100"
-                    leave="ease-in-out duration-300"
-                    leaveFrom="opacity-100"
-                    leaveTo="opacity-0"
-                  >
-                    <div className="absolute left-full top-0 flex w-16 justify-center pt-5">
-                      <button
-                        type="button"
-                        className="-m-2.5 p-2.5"
-                        onClick={() => setSidebarOpen(false)}
-                      >
-                        <span className="sr-only">Close sidebar</span>
-                        <XMarkIcon
-                          className="h-6 w-6 text-white"
-                          aria-hidden="true"
-                        />
-                      </button>
-                    </div>
-                  </Transition.Child>
-                  <div className="flex grow flex-col gap-y-5 overflow-y-auto bg-gray-900 px-6 ring-1 ring-white/10">
-                    <div className="flex h-16 shrink-0 items-center">
-                      <Logo />
-                    </div>
-                    <nav className="flex flex-1 flex-col">
-                      <ul role="list" className="flex flex-1 flex-col gap-y-7">
-                        <li>
-                          <ul role="list" className="-mx-2 space-y-1">
-                            {navigation.map((item) => (
-                              <>
-                                <li key={item.name}>
-                                  <div className="flex">
-                                    <a
-                                      href={item.href}
-                                      className={classNames(
-                                        item.current
-                                          ? "bg-gray-400 dark:bg-gray-800 text-white"
-                                          : "text-gray-400 hover:text-white hover:bg-gray-400 dark:hover:bg-gray-800",
-                                        "flex-1 group flex justify-between gap-x-3 rounded-md p-2 text-sm leading-6 font-semibold"
-                                      )}
-                                    >
-                                      <div className="flex gap-x-3">
-                                        <item.icon
-                                          className="h-6 w-6 shrink-0"
-                                          aria-hidden="true"
-                                        />
-                                        {item.name}
-                                      </div>
-                                      <a
-                                        href={`${item.href}/new`}
-                                        className="flex-none"
-                                      >
-                                        <PlusIcon className="h-6 w-6 shrink-0" />
-                                      </a>
-                                    </a>
-                                  </div>
-                                </li>
-                              </>
-                            ))}
-                          </ul>
-                        </li>
-                        <li className="-mx-6 mt-auto">
-                          <a
-                            href="#"
-                            className="flex items-center gap-x-4 px-6 py-3 text-sm font-semibold leading-6 text-white hover:bg-gray-800"
-                          >
-                            <img
-                              className="h-8 w-8 rounded-full bg-gray-800"
-                              src="https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80"
-                              alt=""
-                            />
-                            <span className="sr-only">Your profile</span>
-                            <span aria-hidden="true">Tom Cook</span>
-                          </a>
-                        </li>
-                      </ul>
-                    </nav>
-                  </div>
-                </Dialog.Panel>
-              </Transition.Child>
-            </div>
-          </Dialog>
-        </Transition.Root>
-
-        {/* Static sidebar for desktop */}
-        <div className="hidden xl:fixed xl:inset-y-0 xl:z-50 xl:flex xl:w-72 xl:flex-col">
-          <div className="flex grow flex-col gap-y-5 overflow-y-auto bg-black/10 px-6 ring-1 ring-white/5">
-            <Logo />
-            <nav className="flex flex-1 flex-col">
-              <ul role="list" className="flex flex-1 flex-col gap-y-7">
-                <li>
-                  <ul role="list" className="-mx-2 space-y-1">
-                    {navigation.map((item) => (
-                      <>
-                        <li key={item.name}>
-                          <div className="flex">
-                            <a
-                              href={item.href}
-                              className={classNames(
-                                item.current
-                                  ? "bg-gray-400 dark:bg-gray-800 text-white"
-                                  : "text-gray-400 hover:text-white hover:bg-gray-400 dark:hover:bg-gray-800",
-                                "flex-1 group flex justify-between gap-x-3 rounded-md p-2 text-sm leading-6 font-semibold"
-                              )}
-                            >
-                              <div className="flex gap-x-3">
-                                <item.icon
-                                  className="h-6 w-6 shrink-0"
-                                  aria-hidden="true"
-                                />
-                                {item.name}
-                              </div>
-                              <a
-                                href={`${item.href}/new`}
-                                className="flex-none"
-                              >
-                                <PlusIcon className="h-6 w-6 shrink-0" />
-                              </a>
-                            </a>
-                          </div>
-                        </li>
-                      </>
-                    ))}
-                  </ul>
-                </li>
-                <li className="-mx-6 mt-auto">
-                  <a
-                    c
-                    href="#"
-                    className="flex items-center gap-x-4 px-6 py-3 text-sm font-semibold leading-6 text-white hover:bg-gray-800"
-                  >
-                    <span aria-hidden="true">Sign In</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
+      <div className="xl:pl-72">
+        {/* Sticky search header */}
+        <div className="sticky top-0 z-40 flex h-16 shrink-0 items-center gap-x-6 border-b dark:border-white/5 dark:bg-gray-900 px-4 sm:px-6 lg:px-8">
+          <div className="flex flex-1 gap-x-4 self-stretch lg:gap-x-6">
+            <form className="flex flex-1" action="#" method="GET">
+              <label htmlFor="search-field" className="sr-only">
+                Search
+              </label>
+              <div className="relative w-full">
+                <MagnifyingGlassIcon
+                  className="pointer-events-none absolute inset-y-0 left-0 h-full w-5 text-gray-500"
+                  aria-hidden="true"
+                />
+                <input
+                  id="search-field"
+                  className="block h-full w-full border-0 bg-transparent py-0 pl-8 pr-0 text-white focus:ring-0 sm:text-sm"
+                  placeholder="Search..."
+                  type="search"
+                  name="search"
+                />
+              </div>
+            </form>
           </div>
         </div>
 
-        <div className="xl:pl-72">
-          {/* Sticky search header */}
-          <div className="sticky top-0 z-40 flex h-16 shrink-0 items-center gap-x-6 border-b dark:border-white/5 dark:bg-gray-900 px-4 sm:px-6 lg:px-8">
-            <button
-              type="button"
-              className="-m-2.5 p-2.5 text-white xl:hidden"
-              onClick={() => setSidebarOpen(true)}
-            >
-              <span className="sr-only">Open sidebar</span>
-              <Bars3Icon className="h-5 w-5" aria-hidden="true" />
-            </button>
-
-            <div className="flex flex-1 gap-x-4 self-stretch lg:gap-x-6">
-              <form className="flex flex-1" action="#" method="GET">
-                <label htmlFor="search-field" className="sr-only">
-                  Search
-                </label>
-                <div className="relative w-full">
-                  <MagnifyingGlassIcon
-                    className="pointer-events-none absolute inset-y-0 left-0 h-full w-5 text-gray-500"
-                    aria-hidden="true"
-                  />
-                  <input
-                    id="search-field"
-                    className="block h-full w-full border-0 bg-transparent py-0 pl-8 pr-0 text-white focus:ring-0 sm:text-sm"
-                    placeholder="Search..."
-                    type="search"
-                    name="search"
-                  />
-                </div>
-              </form>
-              here is where the new content would go.
-            </div>
-          </div>
-
-          <main className="lg:pr-96">
-            <header className="flex items-center justify-between border-b dark:border-white/5 px-4 py-4 sm:px-6 sm:py-6 lg:px-8">
-              <h1 className="text-base font-semibold leading-7 dark:text-white">
-                Fishing Rigs
-              </h1>
-            </header>
-          </main>
-        </div>
+        <main className="lg:pr-96">
+          <header className="flex items-center justify-between border-b dark:border-white/5 px-4 py-4 sm:px-6 sm:py-6 lg:px-8">
+            <h1 className="text-base font-semibold leading-7 dark:text-white">
+              Fishing Rigs
+            </h1>
+          </header>
+          oh, the main content should probably go here.
+        </main>
       </div>
     </>
   );
